refactor: extract shared estimation heuristics into _lib
Move the built-in model-family denominator rules to _lib/heuristics.ts and the
provider tool-payload shapes, JSON-schema readers and OpenAI cookbook formula
to _lib/tool-payloads.ts. Contextimate keeps its config layering and internals
surface on top of the shared slice.

Groundwork for issue #57: cachemire's model-switch forecast must use the same
numbers as contextimate so the family cannot disagree about the same bytes.
