+++ scripts/dev/agent-lint-baseline.json
@@ -6,7 +6,7 @@
     "defaultTsMax": 350,
     "files": {
       "extensions/pi-cachemire/index.ts": 1222,
-      "extensions/pi-contextimate/index.ts": 1792,
+      "extensions/pi-contextimate/index.ts": 1637
       "extensions/pi-traceline/index.ts": 1766,
       "tests/cachemire/economics-and-render.test.ts": 352,
       "tests/contract/pi-internals.test.ts": 432,
